Conflicts:
	projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/BundleDetailActivity.java

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/ui/workorder/v2/WorkOrderCard.java
@@ -742,7 +742,7 @@ public class WorkOrderCard extends RelativeLayout {
 
     private final WithdrawRequestDialog.OnWithdrawListener _withdrawRequestDialog_onWithdraw = new WithdrawRequestDialog.OnWithdrawListener() {
         @Override
-        public void onWithdraw(long workOrderId) {
+        public void onWithdraw(int workOrderId) {
             if (_workOrder.getId() == workOrderId)
                 WorkOrderTracker.onActionButtonEvent(App.get(), _savedSearchTitle + " Saved Search", WorkOrderTracker.ActionButton.WITHDRAW, WorkOrderTracker.Action.WITHDRAW, workOrderId);
         }
